Refactor using CostCalculators
