+++ mgmap/src/main/java/mg/mgmap/activity/mgmap/features/routing/profile/CostCalculator.java
@@ -1,0 +1,5 @@
+package mg.mgmap.activity.mgmap.features.routing.profile;
+
+public abstract class CostCalculator {
+    protected abstract double calcCosts(double dist, double verDist);
+}
