+++ mgmap/src/main/java/mg/mgmap/activity/mgmap/features/routing/profile/MTB_TEST2.java
@@ -7,11 +7,9 @@ import mg.mgmap.generic.graph.WayAttributs;
 
 public class MTB_TEST2 extends GenRoutingProfile {
     public MTB_TEST2( ) {
-        super(8.0, 0.12, 3, 0, -0.27, 2);
+        super(new ProfileCostCalculator(8.0, 0.12, 3, 0, -0.27, 2));
     }
-    public WayAttributs getWayAttributes(Way way){
-        return new WayTagEval(way);  // default implementation, which does not modify costs based on Tags
-    }
+
     @Override
     protected int getIconIdActive() {
         return R.drawable.rp_mtb1;
