+++ mgmap/src/main/java/mg/mgmap/activity/mgmap/features/routing/profile/MTB_TEST3.java
@@ -1,18 +1,20 @@
 package mg.mgmap.activity.mgmap.features.routing.profile;
 
-import org.mapsforge.map.datastore.Way;
-
 import mg.mgmap.R;
 import mg.mgmap.generic.graph.WayAttributs;
 
 public class MTB_TEST3 extends GenRoutingProfile {
 
     public MTB_TEST3( ) {
-        super(8.0, 0.10, 3, 0, -0.27, 2);
+        super(new ProfileCostCalculator(8.0, 0.10, 3, 0, -0.27, 2));
     }
-    public WayAttributs getWayAttributes(Way way){
-        return new WayTagEvalMTB(way);
-   }
+
+    public void refreshWayAttributes(WayAttributs wayAttributs) {
+        if (wayAttributs instanceof WayTagEval ) {
+            WayTagEval wayTagEval = (WayTagEval) wayAttributs;
+            wayTagEval.mCostCalculator = new CostCalculatorMTB(wayTagEval, mProfileCostCalculator);
+        }
+    }
     @Override
     protected int getIconIdActive() {
         return R.drawable.rp_mtb1;
