+++ mgmap/src/main/java/mg/mgmap/activity/mgmap/features/routing/profile/TwoPieceFuncCostCalculator.java
@@ -1,0 +1,37 @@
+package mg.mgmap.activity.mgmap.features.routing.profile;
+
+import static java.lang.Math.abs;
+
+import android.util.Log;
+
+public abstract class TwoPieceFuncCostCalculator extends CostCalculator {
+
+    protected double mUpCosts;
+    protected double mDnCosts;// base costs in m per hm uphill;
+    protected double mUpSlopeLimit; //  up to this slope base Costs
+    protected double mDnSlopeLimit;
+    protected double mUpAddCosts; // relative additional costs per slope increase ( 10 means, that costs double with 10% slope increase )
+    protected double mDnAddCosts;
+
+
+    protected double calcCosts(double dist, double vertDist){
+        if (dist <= 0.0000001 ) {
+            return 0.0001;
+        }
+        double slope = vertDist / dist;
+        if ( abs(vertDist) <= 0.000001 ) Log.e("GenRoutingProfile","Suspicious Slope in calcCosts. Dist:" + dist + " VertDist:" + vertDist + " Slope:" + slope);
+        double cost;
+        if (slope > 0) {
+            if (slope <= mUpSlopeLimit)
+                cost = dist + vertDist * mUpCosts;
+            else
+                cost = dist + vertDist * ( mUpCosts + (slope - mUpSlopeLimit) * mUpAddCosts);
+        } else {
+            if (slope >= mDnSlopeLimit)
+                cost = dist + vertDist * mDnCosts;
+            else
+                return dist + vertDist * ( mDnCosts + (slope - mDnSlopeLimit) * mDnAddCosts);
+        }
+        return cost + 0.0001;
+    }
+}
