--- mgmap/src/main/java/mg/mgmap/activity/mgmap/features/routing/profile/WayAttributes4MTB.java
@@ -1,90 +1,0 @@
-package mg.mgmap.activity.mgmap.features.routing.profile;
-
-import org.mapsforge.core.model.Tag;
-import org.mapsforge.map.datastore.Way;
-
-/** test implementation: To be deleted
- */
-public class WayAttributes4MTB extends WayAttributeTags {
-
-
-    protected double mGenCostFactor;
-    protected double mMultCostFactor = 1;
-
-    public WayAttributes4MTB(Way way){
-        super(way);
-    }
-
-    public void calcCostFactors(){
-        if (accessable) {
-            if ("path".equals(highway)) {
-                if (mtbscale != null) {
-                    switch (mtbscale) {
-                        case "mtbs_0":
-                        case "mtbs_1":
-                            break;
-                        case "mtbs_2":
-                            mGenCostFactor = 1.5;
-                            break;
-                        case "mtbs_3":
-                        default:
-                            mGenCostFactor = 3;
-                    }
-                } else
-                    mGenCostFactor = 3;
-                if (trail_visibility != null) {
-                    switch (trail_visibility) {
-                        case "bad":
-                            mMultCostFactor = 1.5;
-                            break;
-                        case "horrible":
-                        case "no":
-                            mMultCostFactor = 2;
-                            break;
-                    }
-                }
-            } else if ("track".equals(highway)) {
-                mGenCostFactor = 1;
-            } else if ("primary".equals(highway)) {
-                if ("bic_no".equals(bicycle))
-                    accessable = false;
-                else if (cycleway != null)
-                    mGenCostFactor = 2;
-                else
-                    mGenCostFactor = 3;
-            } else if ("secondary".equals(highway)) {
-                if (cycleway != null)
-                    mGenCostFactor = 1.5;
-                else
-                    mGenCostFactor = 2;
-            } else if ("tertiary".equals(highway)) {
-                mGenCostFactor = 1.5;
-            } else if ("steps".equals(highway)) {
-                mGenCostFactor = 4;
-//                setFixUpDistParameter(8);
-//                setFixDownDistParameter(8);
-            } else if ("footway".equals(highway) ) {
-                if ( "bic_yes".equals(bicycle) )
-                    mGenCostFactor = 1;
-                else
-                    mGenCostFactor = 4;
-            }
-            else if ("bic_no".equals(bicycle))
-                mGenCostFactor = 4;
-            else
-                mGenCostFactor = 1;
-        } else
-            mGenCostFactor = 10;
-
-        mGenCostFactor = Math.max( mGenCostFactor * mMultCostFactor, 1);
-
-    }
-
-    public double getGenCostFactor() {
-        return mGenCostFactor;
-    }
-
-    public double getMultCostFactor() {
-        return mMultCostFactor;
-    }
-}
