+++ mgmap/src/main/java/mg/mgmap/activity/mgmap/features/routing/profile/WayTagEval.java
@@ -17,14 +17,8 @@ public class WayTagEval extends WayAttributs {
     protected String trail_visibility = null;
     protected String tracktype = null;
     protected String network = null;
+    protected CostCalculator mCostCalculator;
 
-    protected double mUpCosts;
-    protected double mDnCosts;// base costs in m per hm uphill;
-    protected double mUpSlopeLimit; //  up to this slope base Costs
-    protected double mDnSlopeLimit;
-    protected double mUpAddCosts; // relative additional costs per slope increase ( 10 means, that costs double with 10% slope increase )
-    protected double mDnAddCosts;
-    protected double mGenCostFactor;
 
     public WayTagEval(Way way) {
 
@@ -63,13 +57,5 @@ public class WayTagEval extends WayAttributs {
             accessable = false;
         }
     }
-    protected void calcCostFactors(GenRoutingProfile profile){
-        mUpCosts = profile.mUpCosts;
-        mUpSlopeLimit = profile.mUpSlopeLimit;
-        mUpAddCosts = profile.mUpAddCosts;
-        mDnCosts = profile.mDnCosts;
-        mDnSlopeLimit = profile.mDnSlopeLimit;
-        mDnAddCosts = profile.mDnAddCosts;
-        mGenCostFactor = 1.0;
-    };
+
 }
